move sender node queries to querier
