+++ boot-core/src/daemon/mod.rs
@@ -1,4 +1,5 @@
 pub mod core;
+mod querier;
 pub mod sender;
 pub mod state;
 pub mod tx_resp;
