+++ boot-core/src/daemon/sender.rs
@@ -1,8 +1,12 @@
-use crate::cosmos_modules::{self, auth::BaseAccount};
+use crate::{
+    cosmos_modules::{self, auth::BaseAccount},
+    daemon::querier::DaemonQuerier,
+};
 use cosmrs::{
     bank::MsgSend,
     crypto::secp256k1::SigningKey,
-    tendermint::{chain::Id, Block, Time},
+    proto::traits::Message,
+    tendermint::{chain::Id},
     tx::{self, Fee, Msg, Raw, SignDoc, SignerInfo},
     AccountId, Any, Coin,
 };
@@ -88,7 +92,7 @@ impl Sender<All> {
         msgs: Vec<T>,
         memo: Option<&str>,
     ) -> Result<CosmTxResponse, BootError> {
-        let timeout_height = self.block_height().await? + 10u32;
+        let timeout_height = DaemonQuerier::block_height(self.channel()).await? + 10u64;
         let msgs: Result<Vec<Any>, _> = msgs.into_iter().map(Msg::into_any).collect();
         let msgs = msgs?;
         let gas_denom = self.daemon_state.gas_denom.clone();
@@ -104,7 +108,7 @@ impl Sender<All> {
             ..
         } = self.base_account().await?;
 
-        let tx_body = tx::Body::new(msgs, memo.unwrap_or_default(), timeout_height);
+        let tx_body = tx::Body::new(msgs, memo.unwrap_or_default(), timeout_height as u32);
         let auth_info =
             SignerInfo::single_direct(Some(self.private_key.public_key()), sequence).auth_info(fee);
         let sign_doc = SignDoc::new(
@@ -115,7 +119,7 @@ impl Sender<All> {
         )?;
         let tx_raw = sign_doc.sign(&self.private_key)?;
 
-        let sim_gas_used = self.simulate_tx(tx_raw.to_bytes()?).await?;
+        let sim_gas_used = DaemonQuerier::simulate_tx(self.channel(), tx_raw.to_bytes()?).await?;
 
         log::debug!("{:?}", sim_gas_used);
 
@@ -155,43 +159,8 @@ impl Sender<All> {
         Ok(acc)
     }
 
-    pub async fn simulate_tx(&self, tx_bytes: Vec<u8>) -> Result<u64, BootError> {
-        let _addr = self.pub_addr().unwrap().to_string();
-
-        let mut client = cosmos_modules::tx::service_client::ServiceClient::new(self.channel());
-        #[allow(deprecated)]
-        let resp = client
-            .simulate(cosmos_modules::tx::SimulateRequest { tx: None, tx_bytes })
-            .await?
-            .into_inner();
-        let gas_used = resp.gas_info.unwrap().gas_used;
-        Ok(gas_used)
-    }
-
     pub fn channel(&self) -> Channel {
         self.daemon_state.grpc_channel.clone()
-    }
-
-    async fn latest_block(&self) -> Result<Block, BootError> {
-        let mut client =
-            cosmos_modules::tendermint::service_client::ServiceClient::new(self.channel());
-        #[allow(deprecated)]
-        let resp = client
-            .get_latest_block(cosmos_modules::tendermint::GetLatestBlockRequest {})
-            .await?
-            .into_inner();
-        Ok(Block::try_from(resp.block.unwrap())?)
-    }
-
-    pub async fn block_height(&self) -> Result<u32, BootError> {
-        let block = self.latest_block().await?;
-        Ok(block.header.height.value() as u32)
-    }
-
-    /// Returns the block timestamp (since unix epoch) in nanos
-    pub async fn block_time(&self) -> Result<u128, BootError> {
-        let block = self.latest_block().await?;
-        Ok(block.header.time.duration_since(Time::unix_epoch())?.as_nanos())
     }
 
     async fn broadcast(&self, tx: Raw) -> Result<CosmTxResponse, BootError> {
